Moving SearchFeaturedCard to own file
