+++ src/views/homepage/components/featured-content-grid/featured-content-grid.module.css
@@ -34,101 +34,6 @@
 
 /*
 ***
-* SearchFeaturedCard
-***
-*/
-
-.searchCard {
-	background: url('../../img/search-card-graphic.svg');
-	border-radius: 10px;
-	box-shadow: none;
-	padding-bottom: 18px;
-	padding-left: 24px;
-	padding-right: 24px;
-	padding-top: 20px;
-}
-
-.searchCardText {
-	display: flex;
-	flex-direction: column;
-	gap: 8px;
-	margin-bottom: 20px;
-}
-
-.searchCardCarousel {
-	border-radius: 10px;
-	overflow: hidden;
-	padding-bottom: 38px;
-	padding-top: 38px;
-	position: relative;
-	z-index: 0;
-}
-
-.searchCardCarouselBackground {
-	background: linear-gradient(
-			90deg,
-			rgba(13, 13, 21, 0) -0.96%,
-			rgba(13, 13, 21, 0.42) 20.63%,
-			rgba(13, 13, 21, 0.4) 79.73%,
-			rgba(13, 13, 21, 0) 104.43%
-		),
-		#2a283a40;
-	position: absolute;
-	top: 0;
-	right: 0;
-	left: 0;
-	bottom: 0;
-	z-index: -1;
-}
-
-.searchCardList {
-	align-items: center;
-	display: flex;
-	gap: 20px;
-	list-style: none;
-	margin: 0;
-	padding: 0;
-}
-
-.searchCardListItem {
-	flex-shrink: 0;
-}
-
-.searchCardButtonContainer {
-	align-items: center;
-	display: flex;
-	gap: 24px;
-	max-width: 100%;
-	overflow: hidden;
-	position: relative;
-	scroll-snap-type: x mandatory;
-	scroll-behavior: smooth;
-
-	&::before,
-	&::after {
-		content: '';
-		min-height: 1px;
-		min-width: 300px;
-	}
-}
-
-.searchCardButton {
-	align-items: center;
-	background-color: var(--token-color-foreground-strong);
-	border-radius: 10px;
-	color: var(--token-color-palette-neutral-0);
-	display: flex;
-	gap: 16px;
-	border: none;
-	padding-bottom: 16px;
-	padding-left: 24px;
-	padding-right: 24px;
-	padding-top: 16px;
-	scroll-snap-align: center;
-}
-
-/*
-***
 * Certifications FeaturedCard
 ***
 */
